clone_twitter/user/views.py
from django.shortcuts import render
from rest_framework import status, permissions
from rest_framework.views import Response, APIView
from drf_yasg.utils import swagger_auto_schema
from drf_yasg import openapi

# Create your views here.

class PingPongView(APIView):
    permission_classes = (permissions.AllowAny,)

    swagger_auto_schema(request_body=openapi.Schema(
        type=openapi.TYPE_OBJECT,
        properties={
            'ping': openapi.Schema(type=openapi.TYPE_STRING, description='ping'),
        }
    ))

    def get(self, request):
        return Response(data={'ping': 'pong'}, status=status.HTTP_200_OK)

